# specs/openapi.yaml
openapi: 3.0.3
info:
  title: HouseFlow API
  description: |
    API pour l'application HouseFlow - Suivi d'entretien de maison.

    ## MVP Features
    - Authentification (inscription, connexion, déconnexion)
    - Gestion des maisons (CRUD)
    - Gestion des appareils (CRUD)
    - Gestion des types d'entretien (CRUD)
    - Logging des entretiens effectués
    - Calcul des scores de progression

    ## Conventions
    - **Erreurs** : ProblemDetails (`application/problem+json`, schéma `ErrorResponse`) avec un
      `code` machine quand l'interface doit réagir (voir `ErrorResponse.code`).
    - **403 / 404** : 404 (`not_found`) si la ressource — ou la maison qui la contient — n'existe
      pas ; 403 (`forbidden`) si elle existe mais que l'appelant n'a pas le rôle requis (R5).
    - **Dates calendaires** (date d'entretien, échéance, date d'installation) : un jour du
      calendrier Europe/Paris, transporté comme **minuit UTC de ce jour**
      (`yyyy-MM-ddT00:00:00Z`). Les réponses l'envoient toujours sous cette forme
      (`format: date-time`) ; les requêtes acceptent `yyyy-MM-dd` ou `yyyy-MM-ddT00:00:00Z`
      (`format: date`) — seule la partie date est prise en compte, jamais un instant local.
  version: 1.0.0
  contact:
    name: HouseFlow Team
servers:
  - url: http://localhost:5203/api/v1
    description: Local Development Server
  - url: https://api.houseflow.app/v1
    description: Production Server
...
paths:
...
  /auth/register:
    post:
      operationId: register
      summary: Inscription d'un nouvel utilisateur
      description: |
        Crée un compte utilisateur. **Aucune maison n'est créée** : la première maison est
        créée par l'onboarding (`POST /houses`, écran P05).
        Retourne un token JWT pour authentification immédiate (refresh token en cookie HttpOnly).

        Avec `invitationToken` : l'email doit être celui de l'invitation, et l'invitation est
        acceptée automatiquement après la création du compte (`joinedHouseId` dans la réponse).
        Un token inconnu ou plus utilisable → 400 `invitation_invalid` ; un email différent →
        400 `invitation_email_mismatch`. Dans les deux cas, aucun compte n'est créé.
      tags: [Auth]
      parameters:
        - name: invitationToken
          in: query
          required: false
          description: Token d'invitation (lien `/invitations/{token}`) à accepter après l'inscription
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              firstName: "Jean"
              lastName: "Dupont"
              email: "jean.dupont@example.com"
              password: "MonMotDePasse123!"
              consentAccepted: true
      responses:
        '200':
          description: Inscription réussie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Données invalides, CGU non acceptées, ou invitation invalide (`invitation_invalid`, `invitation_email_mismatch`)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email déjà utilisé (`email_taken`)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      operationId: login
      summary: Connexion
      description: Authentifie un utilisateur et retourne un token JWT.
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: "jean.dupont@example.com"
              password: "MonMotDePasse123!"
      responses:
        '200':
          description: Connexion réussie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: |
            Connexion refusée. `code` :
            - `invalid_credentials` : email ou mot de passe incorrect (un seul message, sans distinguer les deux) ;
            - `account_restricted` : compte sous limitation du traitement (RGPD Art. 18).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      operationId: refreshToken
      summary: Rafraîchir le token
      description: >-
        Échange le refresh token (cookie HttpOnly `refreshToken`) contre un nouveau
        access token. Le refresh token est tourné à chaque appel ; présenter un token
        déjà tourné (hors fenêtre de grâce de 30 s) révoque toute la famille de tokens
        issue du même login (détection de vol).
      tags: [Auth]
      responses:
        '200':
          description: Token rafraîchi
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: |
            Refresh token absent, invalide, expiré ou réutilisé (`invalid_refresh_token`), ou
            compte sous limitation du traitement, RGPD Art. 18 (`account_restricted`).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/logout:
    post:
      operationId: logout
      summary: Déconnexion
      description: Révoque le refresh token actuel.
      tags: [Auth]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Déconnexion réussie (même si le refresh token est absent ou déjà révoqué). Le cookie est effacé.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          description: Non authentifié
  /auth/revoke:
    post:
      operationId: revokeRefreshToken
      summary: Révoquer le refresh token courant
      description: |
        Révoque le refresh token du cookie `refreshToken` et efface le cookie. Contrairement à
        `/auth/logout`, échoue si le cookie est absent ou le token inconnu.
      tags: [Auth]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Token révoqué
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Cookie absent, token inconnu ou déjà révoqué (`invalid_refresh_token`)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Non authentifié

  # ============================================================================
  # HOUSES (US-010 (#168), US-011 (#169), US-012 (#170), US-020 (#171), US-023 (#174), US-024 (#175))
  # ============================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenu via /auth/login ou /auth/register
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Clé API générée depuis les paramètres utilisateur
...
  schemas:
    MessageResponse:
      type: object
      properties:
        message:
          type: string

    # ==========================================================================
    # ERROR
    # ==========================================================================
    ErrorResponse:
      type: object
      description: |
        Erreur au format ProblemDetails (RFC 9457, `application/problem+json`). Le frontend
        branche sur `status` et `code`, jamais sur le texte (`detail` est en anglais, non
        destiné à l'affichage). Les 400 de validation de modèle ajoutent `errors`.
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
        detail:
          type: string
          nullable: true
          description: Message technique (anglais)
        code:
          type: string
          nullable: true
          description: |
            Code machine stable, quand l'interface doit réagir spécifiquement :
            `invalid_credentials`, `account_restricted`, `invalid_refresh_token`, `email_taken`,
            `export_rate_limited`, `invitation_invalid`, `invitation_email_mismatch`,
            `invitation_already_pending`, `invitation_limit_reached`, `already_member`,
            `own_invitation`, `forbidden`, `not_found`.
        errors:
          type: object
          nullable: true
          additionalProperties:
            type: array
            items:
              type: string
          description: Erreurs de validation par champ (400 de validation uniquement)

    # ==========================================================================
    # AUTH
    # ==========================================================================
    RegisterRequest:
      type: object
      required: [firstName, lastName, email, password, consentAccepted]
      properties:
        firstName:
          type: string
          minLength: 1
          maxLength: 100
          example: "Jean"
        lastName:
          type: string
          minLength: 1
          maxLength: 100
          example: "Dupont"
        email:
          type: string
          format: email
          minLength: 1
          example: "jean.dupont@example.com"
        password:
          type: string
          minLength: 8
          pattern: "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[^A-Za-z0-9]).{8,}$"
          description: Minimum 8 caractères, avec au moins une majuscule, une minuscule, un chiffre et un caractère spécial
          example: "MonMotDePasse123!"
        consentAccepted:
          type: boolean
          description: |
            RGPD — l'utilisateur accepte les Conditions générales d'utilisation (base légale :
            exécution du contrat, Art. 6(1)(b)) et reconnaît avoir pris connaissance de la
            politique de confidentialité (information, Art. 13). Ce n'est PAS un consentement
            au sens de l'Art. 7 (le traitement des données de compte repose sur le contrat).
            Doit être `true`, sinon l'inscription est refusée (400). La date et la version de
            la politique acceptée sont enregistrées à titre de preuve (accountability, Art. 5(2)).
          example: true
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          minLength: 1
        password:
          type: string
          minLength: 1
        rememberMe:
          type: boolean
          default: false
          description: >-
            Session persistante : le cookie de refresh token est conservé 365 jours
            (renouvelés à chaque refresh). Sinon, cookie de session (24 h côté serveur),
            supprimé à la fermeture du navigateur.
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT access token
        refreshToken:
          type: string
          description: Token pour rafraîchir l'access token
        expiresIn:
          type: integer
          description: Durée de validité en secondes
        user:
          $ref: '#/components/schemas/User'
        joinedHouseId:
          type: string
          format: uuid
          nullable: true
          description: Inscription avec `invitationToken` uniquement — maison rejointe (P03 → P09)
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        theme:
          type: string
          enum: [light, dark, system]
          default: system
          description: Préférence de thème utilisateur
        language:
          type: string
          enum: [fr, en]
          default: fr
          description: Langue préférée de l'utilisateur
        consentRequired:
          type: boolean
          default: false
          description: |
            RGPD — `true` si l'utilisateur doit (ré)accepter la politique de
            confidentialité / CGU en vigueur (compte créé avant l'introduction du
            consentement, ou nouvelle version de la politique). Le frontend affiche
            alors une bannière de re-consentement (POST /users/me/consent).
        isAdmin:
          type: boolean
          default: false
          description: Administrateur de la plateforme (accès à l'interface d'administration)

    # ==========================================================================
    # HOUSES
    # ==========================================================================
...